% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict_species}
\alias{predict_species}
\title{Predict Species in Image(s)}
\usage{
predict_species(
  model,
  image_paths,
  country = NULL,
  admin1_region = NULL,
  latitude = NULL,
  longitude = NULL,
  batch_size = 50,
  show_progress = TRUE
)
}
\arguments{
\item{model}{A loaded SpeciesNet model object from \code{load_speciesnet}.}

\item{batch_size}{Number of images to process in one batch (default: 50).}

\item{show_progress}{Whether to show a progress bar (default: TRUE).}
}
\value{
A list containing predictions for each image.
}
\description{
Runs species prediction on one or more images.
}
